id correction

+++ database/migrations/2021_07_06_135349_create_api_sensors_table.php
@@ -16,7 +16,7 @@ class CreateApiSensorsTable extends Migration
         Schema::create('api_sensors', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->bigInteger('api_key_id')->unsigned();
-            $table->bigInteger('sensor_id')->unsigned();
+            $table->integer('sensor_id')->unsigned();
 
             $table->foreign('api_key_id')->references('id')->on('api_keys');
             $table->foreign('sensor_id')->references('id')->on('sensors');
